feat(node): Add node flags and reload on change config
